refactor : 패키지 구조 변경 및 Const 상수 추가
