+++ src/main/kotlin/com/yapp/itemfinder/common/Const.kt
@@ -1,0 +1,7 @@
+package com.yapp.itemfinder.common
+
+import java.time.ZoneId
+
+object Const {
+    val KST_ZONE_ID = ZoneId.of("Asia/Seoul");
+}
