+++ src/test/kotlin/com/yapp/itemfinder/domain/service/SampleUserServiceTest.kt
@@ -1,8 +1,8 @@
-package com.yapp.itemfinder.sample.service
+package com.yapp.itemfinder.domain.service
 
-import com.yapp.itemfinder.sample.entity.SampleUser
-import com.yapp.itemfinder.sample.repository.SampleUserRepository
-import com.yapp.itemfinder.sample.service.dto.CreateUserReq
+import com.yapp.itemfinder.domain.entity.SampleUser
+import com.yapp.itemfinder.domain.repository.SampleUserRepository
+import com.yapp.itemfinder.domain.service.dto.CreateUserReq
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
